+++ app/scripts/controllers/register.js
@@ -35,24 +35,25 @@
         light: false,
         pollution: false,
         bme680: false,
-        co2: false
+        co2: false,
+        dps310: false
       },
       serialPort: 'Serial1',
       soilDigitalPort: 'A',
       soundMeterPort: 'B',
       windSpeedPort: 'C',
-      bmePhenomenon: 'tempHumiPress'
+      bmePhenomenon: 'tempHumiPress',
     };
     vm.display_enabled = false;
     vm.wifi = {
       ssid: '',
-      pasword: ''
+      pasword: '',
     };
 
     vm.ttn = {
       devEUI: '',
       appEUI: '',
-      appKey: ''
+      appKey: '',
     };
 
     // vm.radioModel = null;
@@ -62,7 +63,7 @@
     vm.tag = '';
     vm.geolocationError = {
       error: false,
-      message: ''
+      message: '',
     };
     vm.privacy = false;
     vm.completed = false;
@@ -72,29 +73,29 @@
     vm.newSenseBox = {
       name: '',
       exposure: '',
-      location: []
+      location: [],
     };
 
     vm.modelSelected = {
       id: '',
-      name: false
+      name: false,
     };
     vm.extensions = {
       feinstaub: {
-        id: ''
+        id: '',
       },
       soilMoisture: {
         id: '',
-        port: 'A'
+        port: 'A',
       },
       soundLevelMeter: {
         id: '',
-        port: 'B'
+        port: 'B',
       },
       windSpeed: {
         id: '',
-        port: 'C'
-      }
+        port: 'C',
+      },
     };
 
     vm.mqtt = {
@@ -103,7 +104,7 @@
       topic: '',
       messageFormat: '',
       decodeOptions: '',
-      connectionOptions: ''
+      connectionOptions: '',
     };
     vm.ttnEnabled = false;
     vm.validTTNconfig = true;
@@ -113,7 +114,7 @@
       app_id: '',
       dev_id: '',
       decodeOptions: '[]',
-      cayenneLppDecoding: []
+      cayenneLppDecoding: [],
     };
     vm.open = {
       sensebox: false,
@@ -122,7 +123,7 @@
       mqtt: false,
       ttn: false,
       hackair: false,
-      edu: false
+      edu: false,
     };
 
     vm.markers = {};
@@ -130,8 +131,9 @@
     vm.editMarkerInput = {};
     vm.editing = {};
     vm.events = {
-      autolocation: true
+      autolocation: true,
     };
+    vm.showSameAddressWarning = false;
 
     vm.enterEvent = enterEvent;
     vm.add = add;
@@ -150,6 +152,7 @@
     vm.addSensorTemplate = addSensorTemplate;
     vm.generateScript = generateScript;
     vm.compile = compile;
+    vm.selectBox = selectBox;
 
     activate();
 
@@ -199,7 +202,7 @@
         devEUI: vm.ttn.devEUI,
         appEUI: vm.ttn.appEUI,
         appKey: vm.ttn.appKey,
-        display_enabled: vm.display_enabled
+        display_enabled: vm.display_enabled,
       })
         .then(function (response) {
           vm.boxScript = response;
@@ -212,7 +215,7 @@
 
       return AccountService.compileSketch({
         board: 'sensebox-mcu',
-        sketch: vm.boxScript
+        sketch: vm.boxScript,
       })
         .then(function () { })
         .catch(function () { })
@@ -234,9 +237,8 @@
     }
 
     function setStepTitle () {
-      vm.stepTitle = WizardHandler.wizard(
-        'RegistrationWizard'
-      ).currentStepTitle();
+      vm.stepTitle =
+        WizardHandler.wizard('RegistrationWizard').currentStepTitle();
     }
 
     function stepBack () {
@@ -273,12 +275,32 @@
       }
     }
 
+    function selectBox ($item) {
+      centerLatLng($item.boundingbox);
+    }
+
+    // centers a latlng (marker) on the map while reserving space for the sidebar
+    function centerLatLng (latlng) {
+      osemMapData.getMap('map_register').then(function (map) {
+        map.fitBounds(
+          [
+            [latlng[0], latlng[2]],
+            [latlng[1], latlng[3]],
+          ],
+          {
+            paddingTopLeft: [0, 0],
+            animate: false,
+            zoom: 18,
+          }
+        );
+      });
+    }
+
     function enterEvent (keyEvent) {
       var stepNumber = -1;
       if (keyEvent.which === 13) {
-        stepNumber = WizardHandler.wizard(
-          'RegistrationWizard'
-        ).currentStepNumber();
+        stepNumber =
+          WizardHandler.wizard('RegistrationWizard').currentStepNumber();
       }
       switch (stepNumber) {
       case 2:
@@ -308,7 +330,7 @@
         icon: icon,
         title: title,
         unit: unit,
-        sensorType: sensorType
+        sensorType: sensorType,
       };
       vm.sensors.push(sensor);
       vm.sensorIncomplete = false;
@@ -404,6 +426,9 @@
               case 'co2':
                 vm.newSenseBox.sensorTemplates.push('scd30');
                 break;
+              case 'dps310':
+                vm.newSenseBox.sensorTemplates.push('dps310');
+                break;
               }
             }
           }
@@ -448,7 +473,7 @@
           $translate('REGISTRATION_SUCCESS').then(function (msg) {
             var alert = {
               type: 'success',
-              msg: msg
+              msg: msg,
             };
             vm.alerts.push(alert);
             vm.regSuccess = true;
@@ -462,7 +487,7 @@
           $translate('REGISTRATION_FAIL').then(function (msg) {
             var alert = {
               type: 'danger',
-              msg: msg
+              msg: msg,
             };
             vm.alerts.push(alert);
           });
@@ -605,16 +630,27 @@
         unit = 'ppm';
         sensorType = 'SCD30';
         break;
+      case 'DPS310_AIRPRESSURE':
+        icon = 'osem-barometer';
+        title = 'Luftdruck';
+        unit = 'hPa';
+        sensorType = 'DPS310';
+        break;
+      case 'DPS310_TEMPERATURE':
+        icon = 'osem-thermometer';
+        title = 'Temperatur';
+        unit = '°C';
+        sensorType = 'DPS310';
+        break;
       }
 
       return {
         icon: icon,
         title: title,
         unit: unit,
-        sensorType: sensorType
+        sensorType: sensorType,
       };
     }
-
 
     function removeSensorTemplate (template) {
       if (template === '' || template === null || template === undefined) {
@@ -624,7 +660,10 @@
       // Remove specific sensor template
       for (var index = 0; index < vm.sensors.length; index++) {
         var element = vm.sensors[index];
-        if (element.sensorType === template.sensorType && element.title === template.title) {
+        if (
+          element.sensorType === template.sensorType &&
+          element.title === template.title
+        ) {
           vm.sensors.splice(index, 1);
         }
       }
@@ -637,7 +676,6 @@
       });
 
       angular.copy(tempSensors, vm.sensors);
-
     }
 
     function addSensorTemplate (templateName) {
@@ -649,7 +687,10 @@
       var icon = 'circle';
       var color = 'red';
 
-      if (vm.newSenseBox.exposure === 'indoor' || vm.newSenseBox.exposure === 'outdoor') {
+      if (
+        vm.newSenseBox.exposure === 'indoor' ||
+        vm.newSenseBox.exposure === 'outdoor'
+      ) {
         icon = 'cube';
         color = 'green';
       } else if (vm.newSenseBox.exposure === 'mobile') {
@@ -661,33 +702,32 @@
         type: 'awesomeMarker',
         prefix: 'fa',
         icon: icon,
-        markerColor: color
+        markerColor: color,
       });
     }
 
     ////
 
     $scope.$on('osemMapClick.map_register', function (e, args) {
-
       if (Object.keys(vm.markers).length === 0) {
         vm.markers = {
           box: {
             layerName: 'registration',
             latLng: [
               parseFloat(args.latlng.lat.toFixed(6)),
-              parseFloat(args.latlng.lng.toFixed(6))
+              parseFloat(args.latlng.lng.toFixed(6)),
             ],
             lat: parseFloat(args.latlng.lat.toFixed(6)),
             lng: parseFloat(args.latlng.lng.toFixed(6)),
             draggable: true,
-            icon: generateMarkerIcon()
-          }
+            icon: generateMarkerIcon(),
+          },
         };
       } else {
         vm.markers = angular.copy(vm.markers);
         vm.markers.box.latLng = [
           parseFloat(args.latlng.lat.toFixed(6)),
-          parseFloat(args.latlng.lng.toFixed(6))
+          parseFloat(args.latlng.lng.toFixed(6)),
         ];
         vm.markers.box.lat = parseFloat(args.latlng.lat.toFixed(6));
         vm.markers.box.lng = parseFloat(args.latlng.lng.toFixed(6));
@@ -699,7 +739,7 @@
       vm.markers = angular.copy(vm.markers);
       vm.markers.box.latLng = [
         parseFloat(args.target._latlng.lat.toFixed(6)),
-        parseFloat(args.target._latlng.lng.toFixed(6))
+        parseFloat(args.target._latlng.lng.toFixed(6)),
       ];
       vm.markers.box.lat = parseFloat(args.target._latlng.lat.toFixed(6));
       vm.markers.box.lng = parseFloat(args.target._latlng.lng.toFixed(6));
@@ -714,13 +754,13 @@
             layerName: 'registration',
             latLng: [
               parseFloat(args.latlng.lat.toFixed(6)),
-              parseFloat(args.latlng.lng.toFixed(6))
+              parseFloat(args.latlng.lng.toFixed(6)),
             ],
             lat: parseFloat(args.latlng.lat.toFixed(6)),
             lng: parseFloat(args.latlng.lng.toFixed(6)),
             draggable: true,
-            icon: generateMarkerIcon()
-          }
+            icon: generateMarkerIcon(),
+          },
         };
         if (args.latlng.altitude) {
           vm.markers.box.height = parseFloat(args.latlng.altitude.toFixed(2));
@@ -729,7 +769,7 @@
         vm.markers = angular.copy(vm.markers);
         vm.markers.box.latLng = [
           parseFloat(args.latlng.lat.toFixed(6)),
-          parseFloat(args.latlng.lng.toFixed(6))
+          parseFloat(args.latlng.lng.toFixed(6)),
         ];
         vm.markers.box.lat = parseFloat(args.latlng.lat.toFixed(6));
         vm.markers.box.lng = parseFloat(args.latlng.lng.toFixed(6));
@@ -804,6 +844,35 @@
       if (newValue === 'custom') {
         vm.extensions.feinstaub.id = '';
         vm.newModel.connection = null;
+        vm.newModel.sensors = {
+          temp: false,
+          pressure: false,
+          light: false,
+          pollution: false,
+          bme680: false,
+          co2: false,
+        };
+        vm.modelSelected.name = 'Custom';
+
+        vm.invalidHardware = false;
+
+        return;
+      }
+
+      if (newValue === 'edu') {
+        vm.sensorSetup = '';
+        vm.sensors = [];
+        vm.extensions.feinstaub.id = '';
+        vm.newModel.connection = null;
+        vm.newModel.sensors = {
+          temp: false,
+          pressure: false,
+          light: false,
+          pollution: false,
+          bme680: false,
+          co2: false,
+        };
+        vm.modelSelected.name = 'senseBox Edu';
 
         vm.invalidHardware = false;
 
@@ -831,6 +900,9 @@
         addSensorTemplate('BME680_VOC');
       } else if (newValue.co2 && oldValue.co2 === false) {
         addSensorTemplate('scd30_co2');
+      } else if (newValue.dps310 && oldValue.dps310 === false) {
+        addSensorTemplate('DPS310_AIRPRESSURE');
+        addSensorTemplate('DPS310_TEMPERATURE');
       }
 
       // Remove sensor templates
@@ -849,44 +921,71 @@
         removeSensorTemplate(generateSensorTemplate('BME680_VOC'));
       } else if (oldValue.co2 && newValue.co2 === '') {
         removeSensorTemplate(generateSensorTemplate('scd30_co2'));
+      } else if (oldValue.dps310 && newValue.dps310 === '') {
+        removeSensorTemplate(generateSensorTemplate('DPS310_AIRPRESSURE'));
+        removeSensorTemplate(generateSensorTemplate('DPS310_TEMPERATURE'));
+      }
+
+      // Check on change for sensors with same address
+      if ([vm.newModel.sensors.bme680, vm.newModel.sensors.dps310, vm.newModel.sensors.pressure].filter(Boolean).length >= 2) {
+        vm.showSameAddressWarning = true;
+      } else {
+        vm.showSameAddressWarning = false;
       }
     }, true);
 
     // Watch extensions because they also add sensors
-    $scope.$watch('register.extensions', function (newValue, oldValue) {
-      // Add sensor template
-      if (newValue.feinstaub.id !== '') {
-        addSensorTemplate('PM25');
-        addSensorTemplate('PM10');
-      } else if (newValue.soilMoisture.id !== '') {
-        addSensorTemplate('smt50_soilmoisture');
-        addSensorTemplate('smt50_soiltemperature');
-      } else if (newValue.soundLevelMeter.id !== '') {
-        addSensorTemplate('soundlevelmeter');
-      } else if (newValue.windSpeed.id !== '') {
-        addSensorTemplate('windspeed');
-      }
+    $scope.$watch(
+      'register.extensions',
+      function (newValue, oldValue) {
+        // Add sensor template
+        if (newValue.feinstaub.id !== '') {
+          addSensorTemplate('PM25');
+          addSensorTemplate('PM10');
+        } else if (newValue.soilMoisture.id !== '') {
+          addSensorTemplate('smt50_soilmoisture');
+          addSensorTemplate('smt50_soiltemperature');
+        } else if (newValue.soundLevelMeter.id !== '') {
+          addSensorTemplate('soundlevelmeter');
+        } else if (newValue.windSpeed.id !== '') {
+          addSensorTemplate('windspeed');
+        }
 
-      // Remove sensor template
-      if (newValue.feinstaub.id === '' && oldValue.feinstaub.id !== '') {
-        removeSensorTemplate(generateSensorTemplate('PM25'));
-        removeSensorTemplate(generateSensorTemplate('PM10'));
-      } else if (newValue.soilMoisture.id === '' && oldValue.soilMoisture.id !== '') {
-        removeSensorTemplate(generateSensorTemplate('smt50_soilmoisture'));
-        removeSensorTemplate(generateSensorTemplate('smt50_soiltemperature'));
-      } else if (newValue.soundLevelMeter.id === '' && oldValue.soundLevelMeter.id !== '') {
-        removeSensorTemplate(generateSensorTemplate('soundlevelmeter'));
-      } else if (newValue.windSpeed.id === '' && oldValue.windSpeed.id !== '') {
-        removeSensorTemplate(generateSensorTemplate('windspeed'));
-      }
-    }, true);
+        // Remove sensor template
+        if (newValue.feinstaub.id === '' && oldValue.feinstaub.id !== '') {
+          removeSensorTemplate(generateSensorTemplate('PM25'));
+          removeSensorTemplate(generateSensorTemplate('PM10'));
+        } else if (
+          newValue.soilMoisture.id === '' &&
+          oldValue.soilMoisture.id !== ''
+        ) {
+          removeSensorTemplate(generateSensorTemplate('smt50_soilmoisture'));
+          removeSensorTemplate(generateSensorTemplate('smt50_soiltemperature'));
+        } else if (
+          newValue.soundLevelMeter.id === '' &&
+          oldValue.soundLevelMeter.id !== ''
+        ) {
+          removeSensorTemplate(generateSensorTemplate('soundlevelmeter'));
+        } else if (
+          newValue.windSpeed.id === '' &&
+          oldValue.windSpeed.id !== ''
+        ) {
+          removeSensorTemplate(generateSensorTemplate('windspeed'));
+        }
+      },
+      true
+    );
 
     // Watch added sensors if selected model is custom or edu
-    $scope.$watch('register.sensors', function () {
-      if (vm.ttn.profile === 'cayenne-lpp') {
-        updateCayenneDecoding();
-      }
-    }, true);
+    $scope.$watch(
+      'register.sensors',
+      function () {
+        if (vm.ttn.profile === 'cayenne-lpp') {
+          updateCayenneDecoding();
+        }
+      },
+      true
+    );
 
     $scope.$watch('register.newModel.connection', function (newValue) {
       if (newValue === 'Lora') {
@@ -918,13 +1017,21 @@
 
         // Title could be undefined in manual configuration after adding a sensor
         if (sensor.title) {
-          if (new RegExp(tempSubstr.join('|')).test(sensor.title.toLowerCase())) {
+          if (
+            new RegExp(tempSubstr.join('|')).test(sensor.title.toLowerCase())
+          ) {
             decoderGuess = 'temperature';
-          } else if (new RegExp(humiSubstr.join('|')).test(sensor.title.toLowerCase())) {
+          } else if (
+            new RegExp(humiSubstr.join('|')).test(sensor.title.toLowerCase())
+          ) {
             decoderGuess = 'relative_humidity';
-          } else if (new RegExp(pressSubstr.join('|')).test(sensor.title.toLowerCase())) {
+          } else if (
+            new RegExp(pressSubstr.join('|')).test(sensor.title.toLowerCase())
+          ) {
             decoderGuess = 'barometric_pressure';
-          } else if (new RegExp(illuSubstr.join('|')).test(sensor.title.toLowerCase())) {
+          } else if (
+            new RegExp(illuSubstr.join('|')).test(sensor.title.toLowerCase())
+          ) {
             decoderGuess = 'luminosity';
           }
         }
@@ -933,7 +1040,7 @@
           sensor_title: sensor.title,
           sensor_type: sensor.sensorType,
           decoder: decoderGuess,
-          channel: 1
+          channel: 1,
         });
       });
 
@@ -979,69 +1086,69 @@
       }
     });
 
-    $scope.$watchCollection('register.open', function (
-      newAccordion,
-      oldAccordion
-    ) {
-      if (
-        vm.tag === 'edu' &&
-        newAccordion &&
-        oldAccordion &&
-        oldAccordion.edu === true &&
-        newAccordion.edu === false
-      ) {
-        vm.tag = '';
-      }
-      if (newAccordion) {
-        if (newAccordion.custom === true) {
-          vm.modelSelected.id = 'custom';
-        } else if (newAccordion.edu === true) {
-          vm.modelSelected.id = 'edu';
-          if (vm.tag === '') {
-            vm.tag = 'edu';
+    $scope.$watchCollection(
+      'register.open',
+      function (newAccordion, oldAccordion) {
+        if (
+          vm.tag === 'edu' &&
+          newAccordion &&
+          oldAccordion &&
+          oldAccordion.edu === true &&
+          newAccordion.edu === false
+        ) {
+          vm.tag = '';
+        }
+        if (newAccordion) {
+          if (newAccordion.custom === true) {
+            vm.modelSelected.id = 'custom';
+          } else if (newAccordion.edu === true) {
+            vm.modelSelected.id = 'edu';
+            if (vm.tag === '') {
+              vm.tag = 'edu';
+            }
           }
         }
       }
-    });
+    );
 
-    $scope.$watchCollection('register.editMarkerInput.box', function (
-      newValue,
-      oldValue
-    ) {
-      if (
-        newValue &&
-        newValue.lat &&
-        newValue.lng &&
-        !angular.equals(newValue, oldValue)
-      ) {
-        if (Object.keys(vm.markers).length === 0) {
-          vm.markers = {
-            box: {
-              layerName: 'registration',
-              latLng: [
-                parseFloat(newValue.lat.toFixed(6)),
-                parseFloat(newValue.lng.toFixed(6))
-              ],
-              lat: parseFloat(newValue.lat.toFixed(6)),
-              lng: parseFloat(newValue.lng.toFixed(6)),
-              height: newValue.height,
-              draggable: true,
-              icon: generateMarkerIcon()
-            }
-          };
-        } else {
-          vm.markers = angular.copy(vm.markers);
-          (vm.markers.box.latLng = [
-            parseFloat(newValue.lat.toFixed(6)),
-            parseFloat(newValue.lng.toFixed(6))
-          ]),
-          (vm.markers.box.lat = parseFloat(newValue.lat.toFixed(6)));
-          vm.markers.box.lng = parseFloat(newValue.lng.toFixed(6));
-          vm.markers.box.height = newValue.height;
+    $scope.$watchCollection(
+      'register.editMarkerInput.box',
+      function (newValue, oldValue) {
+        if (
+          newValue &&
+          newValue.lat &&
+          newValue.lng &&
+          !angular.equals(newValue, oldValue)
+        ) {
+          if (Object.keys(vm.markers).length === 0) {
+            vm.markers = {
+              box: {
+                layerName: 'registration',
+                latLng: [
+                  parseFloat(newValue.lat.toFixed(6)),
+                  parseFloat(newValue.lng.toFixed(6)),
+                ],
+                lat: parseFloat(newValue.lat.toFixed(6)),
+                lng: parseFloat(newValue.lng.toFixed(6)),
+                height: newValue.height,
+                draggable: true,
+                icon: generateMarkerIcon(),
+              },
+            };
+          } else {
+            vm.markers = angular.copy(vm.markers);
+            (vm.markers.box.latLng = [
+              parseFloat(newValue.lat.toFixed(6)),
+              parseFloat(newValue.lng.toFixed(6)),
+            ]),
+            (vm.markers.box.lat = parseFloat(newValue.lat.toFixed(6)));
+            vm.markers.box.lng = parseFloat(newValue.lng.toFixed(6));
+            vm.markers.box.height = newValue.height;
+          }
+          vm.editMarkerInput = angular.copy(vm.markers);
         }
-        vm.editMarkerInput = angular.copy(vm.markers);
       }
-    });
+    );
 
     $scope.$watchCollection('register.newSenseBox.exposure', function () {
       if (vm.markers.box) {